api/requirements.txt
fastapi[standard]
asyncpg 
sqlalchemy
PyJWT
pydantic